Begin state-machine-based coarse-pass toml parse (7)
Begin light-parsing toml, cover 3 fail states

Some things like:
  - design a state transition graph for doing a coarse
    parse of toml documents.pt
  - because we had a fresh start, we attempted a blue-sky clean-up
    of the “_init.py” pattern (and reproduced pretty much the same
    thing but it now has a new name: “_common_state.py”).
  - against our better judgement, using “.stack” file again
    (instead of digraphs)

Discussion:

This commit is something of snapshot for posterity. in subsequent
commits the support code for this hand-parsing explodes into its
own more thought-out modules; but:

we wanted to capture this moment as it is because it demonstrates
that (for better or worse) it's possible to have a hand-written
state-machine-based parsing without even needing to have any
abstracted functions etc because of how little code may be required,
depending.

but more broadly it bears mentioning discussion what’s going on with
this project on a bigger picture: of all our nine-ish-or-so app-ish
ideas (think the dancing app, see [#431] for a coarse roadmap):
  - most of them need some kind on online storage backend,
    like a database
  - as we’ve discovered, reading from and writing to filesystems
    in any kind of persistent way is something of a challenge
    when you’re on the cloud in containerized things (for reasons)

having fully drank the kool-aide on containerization, we decided we
needed to “conquer” them up-front (as the commits of the last month
or so are testament to; most of them centered on documentation).

(and at first we thought we wanted just docker, then kubernetes, but
now we’re scaling back to just docker again (minikube slogs on our
laptop and we’re realizing it really is overkill for now, and it’s
choking development.)

ANYWAY, this really slowed things down for us because of how daunting,
unsatisfying, non-linear, and cost-associated (and frankly depressing)
was this undertaking of trying to solve for containerized hosing in this
modern age lol.

BUT when we got an ubuntu/apache container running locally (and were
convinced we could push it up to aws using something like ECS or
whatever, we jumped ahead a bit and googled around about golang-
centered docker images and so on.

(around this point is when we got kicked into a bit of a manic episode
(from late night into early morning 01-20), something about blood moon):

as it turns out (and it should be unsurprising), the space of golang
on containers is well trodden ground where you can make small footprint
images with compiled golang (or something)..

this lead us to start seeing golang code and had us considering a
horizontal/vertical approach, were first we get working a simple
docker image of ubuntu/apache working locally, then get it working on
was, then same for golang. suddenly us writing golang felt much closer
then we realized.

the mania came as the convergence of *two* compunctions of ours: one:
we hate learning about a thing without documenting the journey of our
learning (this goes back to learning early rails when we got frustrated
by the decentralized and blog-centric nature of the documentation
ecosystem there; we saw it as a potential missed opportunity at helping
the ecosystem); and two: our other compunction here is the same that is
the driving force behind the “kiss” project.

this is all to say:
  1) we are going to start studying golang like now
  2) all our OCD would be satisfied by taking notes on
     golang using something like kiss-db
  3) kiss-db doesn’t exist at all yet. we wanted to write it in
     golang but see (2).
  4) this was the crowing manic moment: we decided that we would
     do a first pass of kiss-db in python, and then (as we use it
     to learn golang) re-write it piece-by-piece in golang until
     it becomes “self-supporting” as it were.

the 10 or so stack frame items added to “.stack” correspond to 10
little pieces of bright-pink, florescent little task cards sitting
on our desk. we were hoping to get through this “quickly” so we can
get on to golang. let’s just see shall we!..

(times)

  01-03 18:19  start reading getting started finally
  01-05 15:45  skim several of thems things
        15:57  read tutorial taklingbout control costs
        16:38  start reading EC2 user guide
  01-17 01:25  try again to start with docker (ECR)
  01-19 03:13  begin again - learn about running docker locally
        15:10  begin ubuntu apache docker container run locally
  01-20 04:58  considering starting on numbering thing to be crazy
        05:18  zz
        12:24  begin again, reading toml
        15:57  ok finished setting up test environemnt
        16:44  take a break
        17:31  forked out to focus on reading toml file first
                  now take a break
        20:13  begin item
        21:39  finished most of transiton diagram
        21:59  ok finishd working draft of transition diagram
        22:15  first few tests written
        22:25  begin implement state machine (both in general & specific sense)
        22:55  begin testing very crude state machine implementation
        23:10  first test passing
  01-21 00:02  three fail states covered!