Traverse ID’s without validating (33)
Structuralize all emissions

the ‘expression’ pattern of emission is something we do in our sleep;
but especially for the purposing of parsing it falls short: it’s a
smell for the lowest level to be assembling ascii art in its emissions.
rather what is best is that the emission contains structured data, and
the client renders it appropriately.

a side benefit of this is that testing feels much less stupid, because
we aren’t doing hacks with parsing strings to see if the right info
is in the emission.

Maturify the Data model

there is compelling benefit to treating “end of stream” as a token
at some conceptual level for the purpose of modeling the grammar as
a state machine. (indeed in this commit we were able to delete behavior
code for about two cases because we found a more generalized solution.)

however in implementation it’s awkward to do hacks to try to accommodate
this airy concept of “end of stream” in places where what is typically
being processed is a string (line) token.

As such, we needed some special treatment of “end of stream” both at
the model level and in our parsing routine. this why we at first broke
our main parsing loop into several smaller components (currently
functions inside a function; at one point we made a dedicated Parse
class for this before we expanded the responsibility of “parse state”
to hold what is now the “state body”).

(times)
  03:28  begin
  07:36  zz
  15:05  woke
  01:04  (chatted about reggaeton with [redacted] while hacking)
         rewrote the one guy to be structured. the first case.
         now doing second test case
  02:52  ok finished removing all expression
  03:55  now