Add CUD entities roadmap (155)
This is a purely cosmetic commit, but not one without substance:

The entities-level work done in earlier legs was largely a
proof-of-concept to establish that our coarse-parsing would work
for our imagined use-cases.

Then, the last leg was diving into our nitty-gritty of worrying
about comment line proximity while covering CUD attributes.

Now that we have these two bulks of work done, what remains is some
extensive integration of all the work, and re-working/re-fitting.

Having gained knowledge from both of these explorations, we now know
essential details we did not forsee while writing pseudocode, like
that an UPDATE will require that we find the existing entity in the
document *first*, before knowing what the replacement lines will be.

Taking all this into account, we came up with a new unified plan
(actually a regression-friendly testing order) for how to test for
all of CRUD+T for entities in one file; knowing what we know now
about our CUD attributes API.

To accord with this, we have re-numbered and re-ordered our test cases.

While we were at it, we flushed pending renames, so we can make
better sense of how our magnetics fit together.

(The message in the commit before this one was written after it ;))

(times)

  02-06 04:15  break down entity CRUD+T
        09:25  zz
        14:13  resume
  02-07 01:49  finished entities CUD roadmap
        04:17  now