Refactor DELETE/UPDATE entity (lines) (182)
This completes our two-commit refactoring of the *covered* behavior
of CUD entities so that its implementation(s) use “block streams”
rather than parse-actions.

It bears highlighting that this is (intentionally) a “pure refactor”:
we have not added or removed or changed any tests, so we have not
changed any behavior as far as we should care to measure; we have only
changed how we implement the behavior.

A side-by-side comparison of the “old” code (only a few weeks old)
with the new stands alone as testament to the superiority of the new
way. The contrast between the two couldn’t be starker and is probably
worthy of at least one chapter in a book.

(For one thing, the old asset file was 500 lines long. The new one
is 254, and this is with a lot of logical redundancy intentionally
left in.)

We tried so hard under the old way to explain in a lengthy comment why
it appeared so complicated. But this just made it worse because it made
it *seem* complicated — it reinforced our assumption that it was
necessarily complicated, an assumption we hopefully proven false here.

This is what catalyzed the whole refactor:

When the time came to add a feature, this was when our approach really
showed strain. (Indeed in his _Clean Code_, Robert C. Martin suggests
(more or less) that the cost (think _time_) of changing code (e.g adding
features) is a good metric for the “cleanliness” of it.)

Specifically we wanted to add behavior that was by design specified
*against* in the original take: now we want to ensure that the whole
file is valid in terms of the order of items before we save a modified
version of it. (Now as opposed to before, when doing writes we want to
be sure that we aren’t propagating forward any “corruption” (possibly
human-introduced), regardless of the extra overhead from checking for
things like order. Note that for reads, near-term we are likely to
continue to optimize for some combination of execution speed/memory
frugality/code simplicity.)

Amazingly, to add this new feature was especially painful with our
parse-actions dependency-injection setup; despite how new and
overly … coddled the code was.

The thinking was, “why can’t this just be a loop?”. Especially with
python the ability to use iterators here holds a lot of value because
the code can still be read in a straightforward procedural/imperative
style even though in reality it may be executing in a streaming
(“pull”/“start/stop”) manner.

Anyway this two-commit snapshot feels like it might hold other similar
lessons if given closer inspection.

(times)

  02-13 22:43  begin DELETE rewrite
        22:55  pause b.c cafe closing
        23:54  begin again
        11:55  done with DELETE
        23:57  begin UPDATE
  02-14 00:12  done with UPDATE!
        01:02  now