Introduce collections, file-rewrite requests, ID's (199)
We want to believe that we are “coming down the mountain” after that
crazy detour with parse-actions-based CUD implementation. (This is to
say: if there’s anything as difficult as that before we get to “alpha”
(i.e CLI/python), we certainly don’t see it coming.)

In this commit we put a lot of doc work into trying to line up the
the remaining steps so interdependencies link up and nothing gets
stalled or “clogged” due to adding too many new parts at once.
(this is reflected in “.stack” and a graph viz document here.)

Also note we are *already* re-allocating test-case numbers because we
like the new convention so much, we want to see if it’s worth it to try
to fit most of the story into the same number space.

Developmentally the main interesting thing we add here is a dependency
injection of a “filesystem”, which lets us abstract away a future
complexity while making the interesting part much easier to test
(so develop), because while we read from the filesystem we never
actually write to it (we just sort of mock it with something like a spy).

(times)
  02-14 01:58  improve the dependency modeling of the “endgame” graph
        04:05  the endgame dependency graph is much more fleshed out
               (plus stack populated). now, going to sketch a
               specification for schema/collection
        04:59  wrote doc about filetree schema (lots of new sections) zz
        11:36  ok look at dep chart and doc and double check assumptions
               to come up with some cases.
               actually write "implementation decisions"
        13:15  ok wrote a thing about mutable collection architecture
        16:09  ok renumbered to make more space lol
        16:41  ok provisioned numbers for new tests. almost zz then no.
        17:33  ok 7 test cases for delete are named
        18:32  create "collection via directory" asset file
        20:11  passing first interesting test whining about bad ID's
        20:51  completed test about identifier depth
        22:30  stop to go see shigeto et. al
  02-15 02:40  begin again
        03:24  directory not found
               begin file not found
        03:39  file not found done. next: entity not found! woot
        03:44  too tired zz
        10:47  begin again
        11:57  begin writing filesystem spy with no spec!
        16:41  we have made the filesystem spy and it seems to work
               quite well but now we need a nap. (then family evening.)
  02-16 00:40  begin again. got sleepy fast. zz 01:15 till like 04:00
        08:09  maybe try again
        08:33  done with not found integrated with a bunch of stuff
        08:46  begin first real money shot
        09:06  while listening to marvin gay "i want you to want me too"
               on NTS, saw first lines rewritten
        09:49  3 tests sold for that. okay this is epic. zz again
        12:04  ok after second sleep let's get the bred and finish this
               last test of the leg.
        14:35  now