Introduce UPDATE entity in file (200)
we had to triangulate the midway point between two competing forces
while hoping for no leaky abstraction: on the one hand, keeping the
line-based CUD operations as purely line-based, it makes testing
easier and feels like it creates good abstraction boundaries.

on the other hand, the implementation for UPDATE (here we truly
discover) cannot be purely line-based in the most straightforward
sense because the new lines occur as the result of a function of
the existing entity. so for UPDATE we inject a function instead of
passing lines. that's it!

(times)
  02-16 15:03  maybe begin looking at an integrated update
        15:47  about to write the update method on the collection class
        16:44  good progress. update works (not tested) but we broke
               the old way. have to break for dinner
  02-17 00:31  work for 1:29 then zz
        09:41  begin again
        14:15  simplified our CUD entities document
  02-18 07:55  finish by injecting the UPDATE func in the right place
        09:20  passing prob the only test gonna make. zz
        14:18  woke, might mess around and commit this
        18:08  now