Introduce “storage schema” (pure abstraction) (268)
Discussion: of the several big nodes remaining to complete our CLI
(integrate DELETE, integrate UPDATE, edge cases near blank or noent
entities files) we decided to go ahead with realizing the beginnings
of our formal “schema” stuff, because of how many corollary nodes will
depend on it.

more specifically, for CREATE and UPDATE of CUD+RT, if the CLI has no
business knowledge about the “primitive type” of the various logical
fields (note currently there *is no* formal concept of fields, more
on this later), then it will just write every field as strings which
will look dumb and be wrong.

to finally fill in all the placeholders where we wanted a schema, doing
this had some corollaries: now, the defining first thing of a collection
is its schema file. as such, when a collection is not found it is now
because of a missing file instead of a noent directory, so this changes
the behavior (in terms of helpful error messages).

(side bar note, at this moment we are leaving uncovered the case of a
yes-ent schema file but no-ent entities dir. this will be very similar
(or identical) to the test case we just refactored for this commit.
but we anticipate that this case will raise an exception because here
we have simplified emission handling. added an item to the stack.)

a *good* corollary of *this*, in turn, is that we are finally getting
an interesting `errno` from the platform (2 lol) for “file not found”,
so this allowed us to slip in our much desired improved support for
passing the errno from deep inside the emissions all the way up and out
into the CLI client.

a very small change is that, since parsing schema files (in toml) is
seen as more primal than parsing entities files, we have moved the
“one module to rule them all” module that houses where toml is parsed.
(the toml integration is only a few very straightforward lines, but
still we wanted to keep it all in the same module in case we expand
error handling of parsing toml, etc.)

we did an absolutely insane thing with cacheing parsed schemas because
we can’t stand the thought of re-parsing the same schema file over and
over again for each CLI invocation we are testing.. this hackery is
confined to the testing environment yikes

(times)
  04-15 04:49  begin playing with graphing remaining fellows
        05:12  begin tearing out schema
        06:43  refactored most of schema stuff .. zz
        13:12  resume pulling out "to stream of entities file paths"
  04-16 20:56  begin making monitor for CLI after quite some thought.
        22:52  we think we wrote it, move out of cafe now
        00:00  finally writing our first real schema file
        00:48  begin writing this commit message.