Mature-ify parsing engine to accommodate multi-line parsing (282)
Discussion:

Thus far we have a proof-of-concept in a working prototype CLI client.

We established that we can use a plain-text format (like toml, but
could be yaml or other) to do basic store and retrieve, one that can
scale out (probably) to some “medium-size” of data but still be human-
readable, easy to hack and understand, and possibly leverage VCS for
journaling/audit trail.

The next big step, then, is to tilt this into the direction of one
of our real-world use-cases. the main two we have in mind have as a
requirement that we accommodate long-ish strings (maybe not more than
a dozen lines but each line not more than 79 characters, say).

(two use cases: “castle of dust” and something like evernote/blogging.)

Supporting one or both of toml’s multi-line string formats was a
natural choice for this; but as it would work out, introducing this
into our “coarse parse” idiom threw a monkey wrench at our
implementation (as we sort of expected it would).

(this is evidenced by the dramatic change in our [#863] state digram
(grammar), which has gotten a step past simple.)

In the new way, there is still a plain old Finite State Automaton
defined with states with names and (formally) allowed transitions
between the names; but now, the engine calls an associated action
for *every* transition; rather than the particular transitions calling
particular actions in a more hand-written ad-hoc manner.

overall it heralded in some welcome changes:
- no more hacky End-of-Stream None-for-function thing that has to
  come at the end
- actions are no longer responsible for making their associated
  transitions, creating a better separation of concerns (it was
  really bad before, once we tried to introduce different actions)
- hot-swapping the parse actions is great and crazy
- no more nasty “turn on line caching”

It is vital that we mention: this commit doesn’t add support for
multi-line strings per se, it’s just a massive overhaul of the existing
parsing infrastructure so that we *can* introduce this feature.

the main code-impact evident here is how much almost everything had to
change now that our data model is not line-centric but block-centric
(cascading blocks). there was especially a lot of business logic built
around detecting lines with comments. formerly the grammar recognized
these and it was reflected in the produced AST; now we just have
“discretionary blocks” and (for now) we forget the distinction between
a comment and a blank line..

(times - we did the bad thing and cut-pasted without saving and
accidentally hit Ctrl-w in GitX, losing all of them! but it was
SEVENTEEN DAYS. SEVENTEEN DAYS)