Add to GUI minimal functionality from tutorial
Discussion:

This builds a GUI with one text field and one button and one label,
and when the button is pressed, the string in the text field is
processed trivally and the result is displayed. This is almost exactly
per the tutorial that is also documented in this commit.

Again this is commited at this state to serve as a reference point
for near future work to diff against in case there is near future
issues getting this to work, we have more regressionponts to help
narrow the search space. (This abundance of caution is in part because
we don't yet have unit tests for reasons described in the
documentation secton called "The bootstrapping problem".)

(times)
  06:43  begin on minimal tutorial app again
  07:12  now