Demonstrate PythonKit in a GUI
As documented in this commit, we were able to bridge the divide between
swift and python (miraculously), but we spent most of the day trying to
bridge the divide between GUI and non-GUI swift code.

This is a bleeding-edge proof-of-concept save point. the way we bring
the dependency into our GUI code is absolutely gross and not ok, but
doing it the right way was well out of scope:

TO RUN THE GUI AND PROVE THE CONCEPT you need to run the build script
added here, which does a git checkout (OF MASTER HEAD YIKES) (we made
a note of the SHA in the code just now) and copies 4 files into our
PhoGUI module. (these four paths are added to .gitignore in this commit.)

Also since we don’t know how to pull in our own modules (or even make
them really) in swift/Cocoa yet, we moved code (rather than copy-
pasting), breaking (perhaps temporarily) the proof-of-concept CLI
that proves PythonKit works there.

(times)

  07:16  begin watching Tanmay Bakshi youtube video.
  08:26  done with video it was good but he skipped the most important
         part to us which is what is the right way to import the module
  08:32  begin being stumped by how to import. read artsy blog
  08:46  read thoughtbot blog
  08:51  read hispter blog
  08:54  read stack overflow question
  09:52  zz
  16:37  awake and alive. figure out the basics of importing modules
  17:05  realized we don't know the first thing about the module system
          so start skimming the second of 3 tutorials from raywenderlich
  17:29  we had to read Tanmay Bakshi's code off his youtube screen to
         infer how it works - it was deceivingly simpler
  19:11  wrote install script that appears to work
  19:30  installed it the gross way
         after like an hour figured out we had to turn off sandboxing
         (exactly as Tanmay already showed us) but learned how step
         debugging feels. wonder of wonder it worked!
  23:54  now